+++ .gitignore
@@ -1,0 +1,5 @@
+__pycache__/Stack.cpython-312.pyc
+__pycache__/Tree.cpython-312.pyc
+utilities/__pycache__/stack.cpython-312.pyc
+utilities/__pycache__/GameClass.cpython-312.pyc
+utilities/__pycache__/game_class.cpython-312.pyc
